:art: improve structure: refactor RobotHardware
