+++ README.md
@@ -3,3 +3,5 @@
 ## Robot code made for the 2019 FGC “Ocean Opportunities” competition
 
 The robot was produced as a form of training for the new members of Team Brazil. As a basis for inspiration, the robot was designed for the 2019 FGC “Ocean Opportunities” competition challenge. The robot was not built, but the code was tested using simulators.
+
+### Robot Description
